+++ src/Exceptions/NetgsmException.php
@@ -4,6 +4,6 @@ namespace TarfinLabs\Netgsm\Exceptions;
 
 use Exception;
 
-class IncorrectPhoneNumberFormatException extends Exception
+class NetgsmException extends Exception
 {
 }