+++ src/NetgsmReportInterface.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace TarfinLabs\Netgsm;
-
-interface NetgsmReportInterface
-{
-}